requirements.txt
# Python package requirements 

numpy==2.0.0
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.1
seaborn==0.13.2
tensorflow==2.17.0
torch==2.3.1
jupyterlab==4.0.2
PyYAML==6.0
torchvision==0.15.2
tqdm==4.65.0

